requirements.txt
absl-py==0.9.0
astetik==1.9.9
astor==0.8.1
attrs==19.1.0
backcall==0.1.0
backports-abc==0.5
bleach==2.1.2
blis==0.2.4
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chances==0.1.9
chardet==3.0.4
chars2vec==0.1.7
cycler==0.10.0
cymem==1.31.2
Cython==0.27.3
decorator==4.2.1
defusedxml==0.6.0
dill==0.2.8.2
docutils==0.16
emoji==0.5.4
entrypoints==0.2.3
ftfy==4.4.3
gast==0.3.3
geonamescache==1.1.0
google-pasta==0.2.0
graphviz==0.8.2
grpcio==1.27.2
h5py==2.10.0
html5lib==1.0.1
idna==2.7
ijson==2.3
imbalanced-learn==0.3.2
imblearn==0.0
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
jedi==0.11.1
Jinja2==2.10
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.6.3
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-word-char-embd==0.20.0
kerasplotlib==0.1.6
kiwisolver==1.2.0
liac-arff==2.1.1
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.3
msgpack==0.6.1
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.3.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pathlib==1.0.1
patsy==0.5.1
pexpect==4.3.1
pickleshare==0.7.4
plac==0.9.6
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.11.3
ptyprocess==0.5.2
pycparser==2.18
pyemd==0.5.1
Pygments==2.2.0
pyparsing==2.4.7
Pyphen==0.9.4
pyrsistent==0.15.2
python-dateutil==2.6.1
python-Levenshtein==0.12.0
pytz==2017.3
PyYAML==5.3.1
pyzmq==16.0.3
qtconsole==4.3.1
regex==2017.4.5
requests==2.19.1
scikit-learn==0.21.2
seaborn==0.10.1
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
srsly==0.0.7
statistics==1.0.3.5
statsmodels==0.11.1
tabulate==0.8.7
talos==0.6.6
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==4.5.3
tqdm==4.23.4
traitlets==4.3.2
typing==3.6.2
ujson==1.35
Unidecode==1.0.22
urllib3==1.23
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.1.0
wrangle==0.6.7